Edm at no longer ref and various fixes
